pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-sessions"
version = "2.0.0"
description = "Backup and analyze Claude Code conversations from ~/.claude/projects. Incremental backup with markdown, HTML, and structured data export plus comprehensive statistics."
readme = "README.md"
authors = [
    {name = "Dustin Kirby"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Communications :: Chat",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Natural Language :: English",
]
keywords = [
    "export-claude-code-conversations", "claude-sessions",
    "claude-code-export-tool", "backup-claude-code-logs", "save-claude-chat-history",
    "claude-jsonl-to-markdown", "extract-claude-sessions", "claude-code-no-export-button",
    "where-are-claude-code-logs-stored", "claude-terminal-logs", "anthropic-claude-code",
    "search-claude-conversations", "claude-code-logs-location", "~/.claude/projects",
    "export-claude-conversations", "extract-claude-code", "backup-claude-sessions"
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
nlp = ["spacy>=3.0"]

[project.urls]
Homepage = "https://github.com/ZeroSumQuant/claude-sessions"
Documentation = "https://github.com/ZeroSumQuant/claude-sessions#readme"
Repository = "https://github.com/ZeroSumQuant/claude-sessions"
Issues = "https://github.com/ZeroSumQuant/claude-sessions/issues"

[project.scripts]
claude-sessions = "claude_sessions:main"
# Legacy aliases for backwards compatibility
claude-extract = "extract_claude_logs:launch_interactive"
claude-logs = "extract_claude_logs:launch_interactive"
claude-search = "search_cli:main"

[tool.setuptools]
py-modules = [
    "claude_sessions",
    "backup",
    "formatters",
    "stats",
    "prompts",
    "extract_claude_logs",
    "interactive_ui",
    "search_conversations",
    "realtime_search",
    "search_cli"
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]